+++ CHANGELOG.md
@@ -18,6 +18,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v10.6.0 - 2020-11-04
+
+### Changed
+
+- 外部決済サービス認証情報をプロダクト検索から取得するように調整
+- update @chevre/api-nodejs-client
+- update @motionpicture/gmo-service
+- update @pecorino/api-nodejs-client
+
+### Removed
+
+- 管理者としての口座開設サービスを削除
+
 ## v10.5.0 - 2020-10-23
 
 ### Changed